:art: improve code structure
