+++ package.json
@@ -40,12 +40,16 @@
 		"got": "^11.8.2",
 		"level-rocksdb": "^5.0.0",
 		"y-leveldb": "^0.1.0",
-		"yjs": "^13.5.12"
+		"yjs": "^13.5.12",
+		"pouchdb-adapter-leveldb": "^7.2.2",
+		"rxdb": "^10.1.0",
+		"rxjs": "^7.4.0"
 	},
 	"devDependencies": {
 		"@types/cross-spawn": "^6.0.2",
 		"@types/fs-extra": "^9.0.13",
 		"@types/jest": "^27.0.2",
+		"@types/pouchdb-adapter-leveldb": "^6.1.3",
 		"gitmoji-changelog": "^2.2.1",
 		"gitmoji-cli": "^4.7.0",
 		"jest": "^27.2.4",
